fixed some code
still broken, but working on it. i don't have longer than 20 minutes to work on any code project at this point in time so i just come and add and change shit as per without compiling. feel free to test and fix. i don't have the time and this is a side project